Refactor: ResponseCode module-api로 이동
